@@ -1,10 +1,18 @@
 {
-    "private": true,
-    "devDependencies": {
-        "@types/node": "16.18.112",
-        "chalk": "5.1.1",
-        "inline-fixtures": "1.1.0",
-        "jsonpath-plus": "7.2.0",
-        "typescript": "4.9.5"
-    }
+  "private": true,
+  "devDependencies": {
+    "@types/node": "16.18.112",
+    "chalk": "5.1.1",
+    "inline-fixtures": "1.1.0",
+    "jsonpath-plus": "7.2.0",
+    "tslib": "^2.8.1",
+    "typescript": "5.4.5",
+    "@bazel/bazelisk": "latest",
+    "@bazel/buildifier": "latest",
+    "@bazel/ibazel": "latest",
+    "@bazel/typescript": "^3.5.1"
+  },
+  "dependencies": {
+    "c8": "^10.1.2"
+  }
 }
