@@ -1,6 +1,6 @@
-const coverage = require('./coverage')
+import * as coverage from './coverage_ts'
 
 coverage.test(0)
 
 // Disabling the statment below should give your 50% coverage
-//coverage.test(2);
+// coverage.test(2);
